feat: User 메소드 옮기기
